requirements.txt
ipykernel==6.21.3
pandas==1.5.3
pyspark==3.3.2
boto3==1.26.106
python-dotenv==1.0.0